Fix some minor issues in the docs
Fixed broken link in DOE driver doc
Got rid of warning about wrong length underline
Some feature tests has `self.` references from the unit tests…
Cleaned those up and moved them under a common FeaureTest.
Changed filename in doe tests to be consistent with recorder tests (for doc purposes)
Cleaned up some `pylint` warnings in a couple of test scripts
